restructure dirs
